% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/yaml-package.R
\docType{package}
\name{yaml-package}
\alias{yaml}
\alias{yaml-package}
\title{yaml: Methods to Convert R Data to YAML and Back}
\description{
Implements the 'libyaml' 'YAML' 1.1 parser and emitter (\url{https://pyyaml.org/wiki/LibYAML}) for R.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://yaml.r-lib.org}
  \item \url{https://github.com/r-lib/r-yaml/}
  \item Report bugs at \url{https://github.com/lib/r-yaml/issues}
}

}
\author{
\strong{Maintainer}: Hadley Wickham \email{hadley@posit.co} (\href{https://orcid.org/0000-0003-4757-117X}{ORCID})

Authors:
\itemize{
  \item Jeremy Stephens [contributor]
  \item Kirill Simonov
}

Other contributors:
\itemize{
  \item Shawn Garbett \email{shawn.garbett@vumc.org} (\href{https://orcid.org/0000-0003-4079-5621}{ORCID}) [contributor]
  \item Yihui Xie (\href{https://orcid.org/0000-0003-0645-5666}{ORCID}) [contributor]
  \item Zhuoer Dong [contributor]
  \item Jeffrey Horner [contributor]
  \item reikoch [contributor]
  \item Will Beasley (\href{https://orcid.org/0000-0002-5613-5006}{ORCID}) [contributor]
  \item Brendan O'Connor [contributor]
  \item Michael Quinn [contributor]
  \item Charlie Gao [contributor]
  \item Gregory R. Warnes [contributor]
  \item Zhian N. Kamvar [contributor]
}

}
\keyword{internal}
